OOP shittycurses
Added a revised version of the shittycurses module. It uses objects instead of standard list containers, and is a lot more elegant. The previous module will probably not be updated from now on. I will instead work on the ooshittycurses module, and it might eventually replace the old shittycurses module.